[chore] book data class 분리 (#44)
