+++ data/src/main/java/com/readme/android/data/remote/model/response/RecentReadResponse.kt
@@ -7,17 +7,3 @@ data class RecentReadResponse(
     @SerializedName("books")
     val books: List<Book>
 )
-
-data class Book(
-    @SerializedName("author")
-    val author: String,
-    @SerializedName("image")
-    val image: String,
-    @SerializedName("isbn")
-    val isbn: String,
-    @SerializedName("subIsbn")
-    val subIsbn: String,
-    @SerializedName("title")
-    val title: String
-)
-
